Allow Float32 and Float64 types
Remove a few `double`s in favor of `REAL` to make GJK more generic.
Factor out definitions of REAL, DIM, TINY, EPSILON into separate header file,
to facilitate compiling all the distinct versions.
DIM not yet generic because hardcoded 3 in DOT_PRODUCT

+++ deps/cameron_gjk_v2.4/gjk.c
@@ -308,7 +308,7 @@ REAL gjk_distance(
    REAL * fdisp, * rdisp;
    struct simplex_point local_simplex;
    int d, compute_both_witnesses, use_default, first_iteration, max_iterations;
-   double oldsqrd;
+   REAL oldsqrd;
 
    assert( NumVertices( obj1)>0 && NumVertices( obj2)>0 );
